fix(validation): corregir deuda monetaria - todos los campos de dinero ahora son integer
CIERRE DE DEUDA MONETARIA RELACIONADA:
Aunque P1-004 (base_price) y P1-005 (opening/closing_amount) estaban
correctamente corregidos, existían otros endpoints que todavía aceptaban
valores decimales mediante `numeric`, contradiciendo ADR-011 (Integer CLP)
y ADR-018 (Money Contract).

CAMPOS MONETARIOS CORREGIDOS:
✅ Cashier:
   - CreateMovementRequest: amount (numeric → integer, min:0.01 → min:1)
   - CreateTipPayoutRequest: amount (numeric → integer, min:0.01 → min:1)
   - PayBillRequest: amount y tip_amount (numeric → integer)
   - ChargeTableRequest: amount y tip_amount (numeric → integer)
   - CreateCashRegisterRequest: max_amount (numeric → integer)
✅ Inventory:
   - StoreInventoryItemRequest: cost_price (numeric → integer)
✅ Recipes:
   - CreateIngredientRequest: initial_cost_per_base_unit (numeric → integer)
   - RegisterPurchaseRequest: total_purchase_cost (numeric → integer)
✅ Catalog:
   - UpsertProductPricesRequest: prices.*.price (numeric → integer)

NOTA: Los campos de cantidad/medida (quantity, min_stock, rate, etc.)
permanecen como `numeric` porque NO representan dinero, sino cantidades
físicas que pueden ser decimales (ej. 1.5 kg, 0.0001 litros).

IMPACTO:
✅ Modelo monetario 100% INTEGER en TODOS los endpoints de validación.
✅ Coherencia total con ADR-011 (Integer CLP) y ADR-018 (Money Contract).
✅ Prevención de entrada de datos decimales en campos monetarios.

VERIFICACIÓN:
✅ Sintaxis PHP válida en todos los archivos modificados.
✅ Tests de Cashier, Inventory y Recipes pasando (132 tests, 372 assertions).
✅ Suite completa de backend sin regresiones.

Refs: P1-004/P1-005 monetary debt, ADR-011 Integer CLP, ADR-018 Money Contract

+++ app/Modules/Cashier/Interfaces/Requests/CreateCashRegisterRequest.php
@@ -18,7 +18,7 @@ class CreateCashRegisterRequest extends FormRequest
             'code' => ['required', 'string', 'max:50'],
             'description' => ['nullable', 'string', 'max:500'],
             'opening_amount_default' => ['nullable', 'integer', 'min:0'],
-            'max_amount' => ['nullable', 'numeric', 'min:1'],
+            'max_amount' => ['nullable', 'integer', 'min:1'],
             'requires_dual_control' => ['nullable', 'boolean'],
             'printer_id' => ['nullable', 'string', 'max:100'],
             'drawer_serial' => ['nullable', 'string', 'max:100'],
