+++ app/Modules/Inventory/Interfaces/Requests/StoreInventoryItemRequest.php
@@ -18,7 +18,7 @@ class StoreInventoryItemRequest extends FormRequest
             'name_translations' => ['required', 'array'],
             'name_translations.es' => ['required', 'string', 'max:255'],
             'unit' => ['required', 'string', 'in:unit,kg,lt'],
-            'cost_price' => ['nullable', 'numeric', 'min:0'],
+            'cost_price' => ['nullable', 'integer', 'min:0'],
             'min_stock' => ['nullable', 'numeric', 'min:0'],
             'is_active' => ['nullable', 'boolean'],
         ];
